refactor: searchRoute 이동 판별 책임 기물에 위임
