+++ src/main/java/domain/piece/Chariot.java
@@ -25,8 +25,4 @@ public class Chariot extends ActivePiece {
         return source.makeColStraightRoute(target);
     }
 
-    @Override
-    public boolean isCannon() {
-        return false;
-    }
 }
