+++ src/main/java/domain/piece/General.java
@@ -29,9 +29,4 @@ public class General extends ActivePiece {
     public List<Position> searchRoute(Position source, Position target) {
         return List.of();
     }
-
-    @Override
-    public boolean isCannon() {
-        return false;
-    }
 }
