+++ src/main/java/domain/piece/Horse.java
@@ -43,8 +43,4 @@ public class Horse extends ActivePiece {
         return List.of(source.addPosition(-1, 0));
     }
 
-    @Override
-    public boolean isCannon() {
-        return false;
-    }
 }
