+++ src/main/java/domain/piece/Soldier.java
@@ -36,9 +36,4 @@ public class Soldier extends ActivePiece {
     public List<Position> searchRoute(Position source, Position target) {
         return List.of();
     }
-
-    @Override
-    public boolean isCannon() {
-        return false;
-    }
 }
